Homepagepaneにpurchaseをもたせた
